#S06P12A305-48 Refator: RequestBookReviewDto 필드 타입 수정

+++ backend/src/main/java/com/ssafy/bjbj/api/bookinfo/dto/RequestBookReviewDto.java
@@ -13,7 +13,7 @@ public class RequestBookReviewDto {
 
     private Long memberSeq;
 
-    private Double starRating;
+    private Integer starRating;
 
     private String summary;
 }
